NO-JIRA Prevent maven-source-plugin being executed twice
With the current Maven config, the maven-source-plugin is executed twice, this results in the sources-jars being attached twice which leads the deploy goal in trying to upload the same sources jars twice to the repository. This slows down the build process and also fails on repositories that disallow updating of existing artifacts.

Additionally the "jar" goal is now replaced with "jar-no-fork" which prevents the maven life cycle being executed twice too.