NO-JIRA Fix description of DiskStoreUsage
In 73c4e399d9c2e8401f2034ffb0c5e42b4a70b053 a description is added to DiskStoreUsage. It incorrectly describes the diskStoreUsage as a percentage. This commit changes it to a fraction which it is (also before the description change). A percentage would be better, since MaxDiskUsage is also specified as percentage.