ARTEMIS-3270: move enforcer execution config so it overrides parent execution
The change in 90101f5b5435922fd0310ec8abb50d186245fcce / #3595 didnt work
as expected since the existing enforcer check already wasnt working. It
isnt overriding the apache parent, which is checking for 3.0.x. Moving
the execution into the build element, alongside the java version check,
allows it to replace the parent execution and enforce 3.5.0+ is used.