NO-JIRA Fixing Thread Leakage on test from UDP thread
When running a UDP connection factory you have to either keep it running, or close it so the UDP thread is closed.
this is an issue for the testsuite as we validate for leaked threads. This needs to be fixed on the test.

(cherry picked from commit 6a3df6412baee054a7acc2ca8dc99d5560dcbbfa)