ARTEMIS-3886 fix CLI operation retry
Commit f8b045bd2d95ec28c70c4d64c85400b28d8baef5 broke the operation
retry as it introduced local variables named "user" and "password" which
overried the class-level variables of the same name. Therefore, when the
user re-enters the username and password on the command-line those
values won't actually be used when attempting to reconnect.

(cherry picked from commit 597953b6e22762518c362d484d1a837ea4630154)