ARTEMIS-4170 remove redundant queue creation for OpenWire
o.a.a.a.c.p.o.a.AMQConsumer#init will *always* try to create a core
queue when creating a consumer for a JMS queue. However, this is
already done in o.a.a.a.c.p.o.a.AMQSession#createConsumer.

(cherry picked from commit 7fe1a20909e40a0b69db174e7340c740b3ebae49)