ARTEMIS-4437: move plugin into its profile, avoid pre-build snapshot resolve.
The prior change from 90ac1e6464f9388afbb6d0fa38770b8509e8c070 in #4629 causes
the root pom build to resolve the module snapshots and can result in their
download, even though they are in the same reactor to build. This typically
happens daily on a developers machine, and on every build of CI jobs
(depending on prior mvn install's and repo caching being done).

Rather than the profile flipping a proeprty to stop the plugin skipping, the
plugin should be in the profile so it is only active if the profile is.

(cherry picked from commit 260f774c192b35eeb9165c5505eae37e91056776)