ARTEMIS-4476 Client Failures Soak Test
I was not able to reproduce the actual issue here, but I heavily used this test during debugging.
This will not serve as a reproducer to the Ghost consumer issue, but this is a valid test.

(cherry picked from commit b041f2cde2fe7d8c9d74708ceef53501b603e88b)