ARTEMIS-5214 Removing invalid test
This test was invalid from day one but we would accept eventual failures back then.
As we now want to be more strict about failures it is time for this test to go.

We would need to implement XA and a more complex test for this. as of now the test is invalid and it needs to go.

(cherry picked from commit 76a1ccf086cc8c9fa93116ae53b42b4da5d80e1b)