ARTEMIS-5232 Removing SharedEventLoopGroupTest
This test does not make sense much sense. The barrier does not seem to always work
It is sleeping for 11 seconds for no reason. I didn't want to accept it when it was committed in the first place
so i'm removing it

(cherry picked from commit df6f8be66925ddb609aab297c021c467037b6bb3)