ARTEMIS-5232 Removing RedeliveryPolicyTest::testRepeatedRedeliveryServerSessionNoCommit
This test makes no sense to me. This Wait.assertEquals seems to be doing just a regular sleep.

If I actually properly fixed the test with assertTrue on this Wait clause the test would always fail. Meaning the test is completely broken.

Since this test is not actually testing anything I'm just removing it as other tests are already validating redeliveries