ARTEMIS-5107 using wrong value in ReplicationStartSyncMessage ctor
The incorrect value has always been used for the `beforeTwoEighteen`
variable. However, this is not actually a problem because the
`beforeTwoEighteen` variable is not necessary. It's only job is to
prevent newer versions from sending extra data to older versions.
However, older version will simply ignore the extra data which means
the `beforeTwoEighteen` variable can be removed completely.

This same compatibility pattern is used in many places for the Core
protocol.

The tests added with the original fix successfully reproduced the
original problem and those tests still pass even with this variable
removed. Also, keep in mind that `decodeRest` is still checking the
version so that it doesn't try to read data that doesn't exist from an
older version.

(cherry picked from commit 3101ac59e09c8722f5afc378162a07f518fc7fe9)