ARTEMIS-5215 Credit handler reset in progress flag when stopped
When a pull mode federation consumer is stopped due to demand being
removed and the credit tup-up handler is also awaiting the Queue backlog
to clear in order to grant a new batch of credit it might exit that
cycle an leave the in-progress flag set to true. Currently this likely
won't trigger a stuck consumer but if the code was to be altered to hold
open a link for some period of time before fully closing in order to
avoid needless attach / detach cycles then it would be possible for the
credit replenishment to get stuck because the previous attempt left the
in-progress flag set to true.

(cherry picked from commit 1633b8bf7e605a7893d97ec8c8d6ff1ba7aac830)