ARTEMIS-5383 performance issues with HierarchicalRepositoryChangeListener
This commit improves performance specifically for removing listeners
from o.a.a.a.c.s.i.HierarchicalObjectRepository. It was previously using
an ArrayList and the cost to remove an Object is O(n) on average
whereas the same operation on a HashSet is O(1) on average assuming good
hash values. Furthermore, the list was protected using a

Switching to a ConcurrentHashSet required a few changes to the listeners
themselves to ensure they have good hashCodes, etc.

This commit relies on existing tests for validation.

(cherry picked from commit c7d0ba7e35ccf4bbdda60a5e59c7a37a752d6dd5)