ARTEMIS-5485 Enforce the queue address before deploying
The queue configs loaded from broker properties use the queue name as
the default value for the address if it is not set. Enforcing the queue
address with the parent address before deploying them removes the
requirement to set the address when it doesn't match the queue name.

(cherry picked from commit ec642bad0acdc1b6980fadbb85d0afdf3752fd50)