ARTEMIS-5469 reformulate previous solution
The previously committed solution to this problem caused
`testPacketIdGeneratorNonCleanSession` in `o.a.a.a.t.i.m.MQTTTest` to
fail. The problem with that solution is that it cleared out too much
state.

This commit fixes that by only clearing out the send quota and leaving
the other state in tact.

(cherry picked from commit ac202810f965b81cf7a0e9fd6a8cc981a16769c6)