ARTEMIS-4753 Fixing PageCleanupWhileReplicaCatchupTest intermittent failures
I had investigated the intermittent failures of this test as part of the
races in paging. It turned out to be a test issue (replication could be
started and stopped multiple times within the life cycle of the
connection).

I added barriers to do it properly.

(cherry picked from commit 68d7dbf9eeefad7baf736a23f3d35912f89a26b6)