ARTEMIS-1759 classloader leaks in Core and JMS Client
Remove fields that hold stack-traces in a few classes. These are mostly
unused and when they are used it is only for logging. The upside of
having additional logging is not worth the risk of leaks and eventual
OOME.

(cherry picked from commit e86ea0f622f059784aa9a66fd4fd15874ac660e6)